+++ components/header.vue
@@ -24,12 +24,14 @@
 
         <VMenu placement="bottom-end" :distance="13" strategy="fixed" container=".c-menu-nav" class="c-menu-nav__item c-menu-more">
           <button type="button" class="c-button c-menu-nav__item-button u-button-reset c-button--center-icon" aria-label="Website Menu Navigation">
-            <span v-show="$device.isDesktop" class="u-icon-untouchable u-icon-wrapper c-button--center-icon">
-              <Menu />
-            </span>
-            <span v-show="$device.isMobileOrTablet" class="u-icon-untouchable u-icon-wrapper c-button--center-icon">
-              <MoreVertical />
-            </span>
+            <transition name="fast" mode="out-in">
+              <span v-if="$device.isDesktop" key="desktop" class="u-icon-untouchable u-icon-wrapper c-button--center-icon">
+                <Menu />
+              </span>
+              <span v-if="$device.isMobileOrTablet" key="mobile" class="u-icon-untouchable u-icon-wrapper c-button--center-icon">
+                <MoreVertical />
+              </span>
+            </transition>
           </button>
 
           <template #popper>
@@ -40,7 +42,7 @@
                   Zufälliges Wort
                 </NuxtLink>
               </li>
-              <li v-for="(item, index) in menuItems" :key="index" class="c-menu-more__item">
+              <li v-for="(item, index) in menuItems" :key="index" class="c-menu-more__item" :class="{ 'is-split': item.title === 'Impressum' }">
                 <NuxtLink :to="item.link">
                   {{ item.title }}
                 </NuxtLink>