+++ CHANGELOG.md
@@ -2,6 +2,122 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.15.0...v1.16.0) (2022-07-16)
+
+
+### Features
+
+* **robots.txt:** add sitemap URL ([9cb0b64](https://github.com/felix-berlin/berliner-schnauze/commit/9cb0b6427017d48819a1b3dc12952bf598b67e17))
+
+# [1.15.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.6...v1.15.0) (2022-07-16)
+
+
+### Bug Fixes
+
+* **button:** wrong color on active state ([c6d6d63](https://github.com/felix-berlin/berliner-schnauze/commit/c6d6d63f82437280987a8bbd5ee4158cb387dec0))
+* **color mode button:** remove transition ([04eb60d](https://github.com/felix-berlin/berliner-schnauze/commit/04eb60d62bd91e22aac4b0793f873f6b754a1502))
+* **color-mode-button:** Mismatching childNodes unsing transition-group ([02f2581](https://github.com/felix-berlin/berliner-schnauze/commit/02f2581daf1d63a03b6770144314d9a8f69cf561))
+
+
+### Features
+
+* **index:** removed desktop letter filter ([dc2be6c](https://github.com/felix-berlin/berliner-schnauze/commit/dc2be6c0fe8a219a4865b73c49ddf4109435ed23))
+* **search words:** aria label reacts on state ([a5a826c](https://github.com/felix-berlin/berliner-schnauze/commit/a5a826c856d7c420b27797f3b770967f32a0f5e8))
+* **sort word direction toggle:** add to index + styling ([b85138d](https://github.com/felix-berlin/berliner-schnauze/commit/b85138d951f3d0d4d66d7d4b64879958bc119fc4))
+* **sort word direction:** add sort function to the general filter function ([f415d2d](https://github.com/felix-berlin/berliner-schnauze/commit/f415d2d9c8a0b635df7f76a25917042c4fa20c67))
+
+## [1.14.6](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.5...v1.14.6) (2022-07-15)
+
+
+### Performance Improvements
+
+* **color-mode-button:** render only the used button ([49def3f](https://github.com/felix-berlin/berliner-schnauze/commit/49def3fad312888e6df9bacaf9ad8e440f25ee23))
+* reduced dom size ([fc8adae](https://github.com/felix-berlin/berliner-schnauze/commit/fc8adae86e812d84d8b24a0867dbc00fcacc8f23))
+
+## [1.14.5](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.4...v1.14.5) (2022-07-08)
+
+
+### Bug Fixes
+
+* filter dropdown text color ([d37e9ea](https://github.com/felix-berlin/berliner-schnauze/commit/d37e9ea77aa62294c4e0d74b1099d506fa716008))
+
+## [1.14.4](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.3...v1.14.4) (2022-07-08)
+
+
+### Bug Fixes
+
+* word of the day loading status (4) ([f5c1adf](https://github.com/felix-berlin/berliner-schnauze/commit/f5c1adf6f981dc5a6ada0610bd0b652e02eec0c1))
+
+## [1.14.3](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.2...v1.14.3) (2022-07-08)
+
+
+### Bug Fixes
+
+* word of the day loading status (3) ([89d6548](https://github.com/felix-berlin/berliner-schnauze/commit/89d65486f64219d8ec20985fc49c05943006f0aa))
+
+## [1.14.2](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.1...v1.14.2) (2022-07-08)
+
+
+### Bug Fixes
+
+* woftd status ([36ae0d0](https://github.com/felix-berlin/berliner-schnauze/commit/36ae0d0eca8c20d9f1ff04f180c8e1a83d05b502))
+* word of the day loading status (2) ([b13ed6b](https://github.com/felix-berlin/berliner-schnauze/commit/b13ed6b45d06e77016a6ea8797615ef68fd32727))
+
+## [1.14.1](https://github.com/felix-berlin/berliner-schnauze/compare/v1.14.0...v1.14.1) (2022-07-08)
+
+
+### Bug Fixes
+
+* word of the day loading status ([25ae0cd](https://github.com/felix-berlin/berliner-schnauze/commit/25ae0cd4859e7a0faefa3c29f85bbcc47802a20c))
+
+# [1.14.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.13.1...v1.14.0) (2022-07-08)
+
+
+### Bug Fixes
+
+* **elements:** dark mode color contrast tooltip color ([57d192c](https://github.com/felix-berlin/berliner-schnauze/commit/57d192c52fac259c6585814feb188144210a8faf))
+
+
+### Features
+
+* **elements:** add crown icon to the word of the day ([a532a1f](https://github.com/felix-berlin/berliner-schnauze/commit/a532a1f9ac2e4baabd93b6efb9f31617bf357d65))
+
+## [1.13.1](https://github.com/felix-berlin/berliner-schnauze/compare/v1.13.0...v1.13.1) (2022-07-02)
+
+
+### Bug Fixes
+
+* date format ([5cc041f](https://github.com/felix-berlin/berliner-schnauze/commit/5cc041f3fbcb7141c5b52719cba7fb29492916ca))
+
+# [1.13.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.12.0...v1.13.0) (2022-07-01)
+
+
+### Bug Fixes
+
+* date format for safari ([f6fc2ab](https://github.com/felix-berlin/berliner-schnauze/commit/f6fc2abb67fd0d252221986bcffa980e4059c8ab))
+* **filter dropdown:** removed hide delay ([cb72ce8](https://github.com/felix-berlin/berliner-schnauze/commit/cb72ce87e825c8ce6817acf5e0ec2051ee595ba4))
+* **single word:** learn more link spacing top ([21e6b17](https://github.com/felix-berlin/berliner-schnauze/commit/21e6b17a9435ca19f53cfa89d74f20aca4d22350))
+
+
+### Features
+
+* better word of the day tooltip position ([18c3393](https://github.com/felix-berlin/berliner-schnauze/commit/18c3393323d2808ba229d4dab341d3fcdca1ca18))
+
+# [1.12.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.11.0...v1.12.0) (2022-07-01)
+
+
+### Features
+
+* **nav:** split menu nav in half ([799819e](https://github.com/felix-berlin/berliner-schnauze/commit/799819e01e4bc8dfc6a503b4e3a127a9ba2bad97))
+* **seo:** create router middleware ([65ecbe4](https://github.com/felix-berlin/berliner-schnauze/commit/65ecbe4be04c5da0d8e0f76a8eab0279e7f8c2cd))
+
+# [1.11.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.10.0...v1.11.0) (2022-06-29)
+
+
+### Features
+
+* add error page ([a4f626a](https://github.com/felix-berlin/berliner-schnauze/commit/a4f626a277e17c805005c3b97f0731e2e2c1c6d2))
+
 # [1.10.0](https://github.com/felix-berlin/berliner-schnauze/compare/v1.9.2...v1.10.0) (2022-06-27)